modules/requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-openai
chromadb
openai